pyproject.toml
[project]
name = "upstageailab-ir-competition-ir-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.13"
dependencies = [
    "accelerate>=1.10.1",
    "elasticsearch==8.8.0",
    "google-genai>=1.36.0",
    "hydra-core>=1.3.2",
    "langchain-google-genai>=1.0.10",
    "langchain-upstage>=0.1.7",
    "openai>=1.30.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=2.3.0",
    "transformers>=4.56.1",
]
